requirements.txt
certifi==2018.1.18
chardet==3.0.4
get==1.0.0
idna==2.6
mecab-python3==0.7
post==1.0.0
public==1.0.0
query-string==1.0.0
request==1.0.0
requests==2.18.4
urllib3==1.22
